cart

+++ sick-fits/backend/src/resolvers/Mutation.js
@@ -3,6 +3,7 @@ const jwt = require('jsonwebtoken');
 const { randomBytes } = require('crypto');
 const { promisify } = require('util');
 const { transport, makeEmail } = require('../mail');
+const { hasPermission } = require('../utils');
 
 const Mutations = {
 	async createItem(parent, args, ctx, info) {
@@ -38,7 +39,14 @@ const Mutations = {
 
 	async deleteItem(parent, args, ctx, info) {
 		const where = { id: args.id };
-		const item = await ctx.db.query.item({ where }, `{ id title }`);
+		const item = await ctx.db.query.item({ where }, `{ id title user { id }}`);
+
+		const ownsItem = item.user.id === ctx.request.userId;
+		const hasPermissions = ctx.request.user.permissions.some((permission) => ['ADMIN', 'DELETE'].includes(permission));
+
+		if (!ownsItem && !hasPermissions) {
+			throw new Error("You don't have permissions to do that");
+		}
 
 		return ctx.db.mutation.deleteItem({ where }, info);
 	},
@@ -157,6 +165,88 @@ const Mutations = {
 
 		return updatedUser;
 	},
+	async updatePermissions(parent, args, ctx, info) {
+		if (!ctx.request.userId) {
+			throw new Error('You must be logged in to do that!');
+		}
+
+		const currentUser = await ctx.db.query.user(
+			{
+				where: {
+					id: ctx.request.userId,
+				},
+			},
+			info
+		);
+
+		hasPermission(currentUser, ['ADMIN', 'PERMISSIONUPDATE']);
+
+		return ctx.db.mutation.updateUser(
+			{
+				data: {
+					permissions: {
+						set: args.permissions,
+					},
+				},
+				where: {
+					id: args.userId,
+				},
+			},
+			info
+		);
+	},
+	async addToCart(parent, args, ctx, info) {
+		const { userId } = ctx.request;
+
+		if (!userId) {
+			throw new Error('You must be logged in to do that!');
+		}
+
+		const [existingCartItem] = await ctx.db.query.cartItems({
+			where: {
+				user: { id: userId },
+				item: { id: args.id },
+			}
+		});
+
+		if (existingCartItem) {
+			return ctx.db.mutation.updateCartItem({
+				where: { id: existingCartItem.id },
+				data: { quantity: existingCartItem.quantity + 1 }
+			}, info);
+		}
+
+		return ctx.db.mutation.createCartItem({
+			data: {
+				user: {
+					connect: { id: userId },
+				},
+				item: {
+					connect: { id: args.id },
+				},
+			}
+		}, info);
+	},
+
+	async removeFromCart(parent, args, ctx, info) {
+		const cartItem = await ctx.db.query.cartItem({
+			where: {
+				id: args.id
+			}
+		}, `{ id, user { id }}`);
+
+		if (!cartItem) throw new Error('No Cart Item Found');
+
+		if (cartItem.user.id !== ctx.request.userId) {
+			throw new Error('Cheating!');
+		}
+
+		return ctx.db.mutation.deleteCartItem({
+			where:
+				{ id: args.id },
+			info
+		});
+	}
 };
 
 module.exports = Mutations;
